feat: Update translations for English and Portuguese locales
The code changes include updating the translations for the English and Portuguese locales. This ensures that the application displays the correct translations for various phrases and messages. The translations were added for new phrases such as "Welcome", "Password", "enterPassword", "Data creation stopped", "The maximum value must be greater than the minimum value.", and "User is not active, please contact the administrator". Additionally, the translations for existing phrases like "Temperature" and "Patient not found" were also updated.